meta-python: Add stopit
stopit is a library to raise asynchronous exceptions in other threads, control
the timeout of blocks or callables with two context managers and two decorators.

The license isn't shipped in the pypi package, so it was extracted from the
github repo at [1].

[1] https://github.com/glenfant/stopit/blob/master/LICENSE

Signed-off-by: Adrian Zaharia <Adrian.Zaharia@windriver.com>
Signed-off-by: Khem Raj <raj.khem@gmail.com>

Issue: LINUXEXEC-30287

This was put into kirkstone but not mickledore and added
to master in June 2023.

(upstream-master rev: 586a4bfbacc7ca26bc5220be913b4f2876c94960)

Signed-off-by: Joe Slater <joe.slater@windriver.com>