frr: Security fix CVE-2023-3748
Issue: LIN1023-1416

CVE-2023-3748:
A flaw was found in FRRouting when parsing certain babeld unicast hello
messages that are intended to be ignored. This issue may allow an
attacker to send specially crafted hello messages with the unicast flag
set, the interval field set to 0, or any TLV that contains a sub-TLV
with the Mandatory flag set to enter an infinite loop and cause a denial
of service.

Reference:
https://nvd.nist.gov/vuln/detail/CVE-2023-3748

Patch from:
https://github.com/FRRouting/frr/commit/ae1e0e1fed77716bc06f181ad68c4433fb5523d0

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-devel on 20230828

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>