python3-pillow: Fix CVE-2023-44271
Issue: LIN1023-2571

An issue was discovered in Pillow before 10.0.0. It is a Denial of
Service that uncontrollably allocates memory to process a given task,
potentially causing a service to crash by having it run out of memory.
This occurs for truetype in ImageFont when textlength in an ImageDraw
instance operates on a long text argument.

References:
https://nvd.nist.gov/vuln/detail/CVE-2023-44271
https://security-tracker.debian.org/tracker/CVE-2023-44271

(LOCAL REV: NOT UPSTREAM) -- Not applicable, upstream out of maintenance

Signed-off-by: Narpat Mali <narpat.mali@windriver.com>