python3-django: fix for CVE-2024-24680
Issue: LIN1023-3618

CVE-2024-24680:
An issue was discovered in Django 3.2 before 3.2.24, 4.2 before 4.2.10,
and Django 5.0 before 5.0.2. The intcomma template filter was subject
to a potential denial-of-service attack when used with very long strings.

Since, there is no ptest available for python3-django so have not tested
the patch changes at runtime.

References:
https://nvd.nist.gov/vuln/detail/CVE-2024-24680
https://security-tracker.debian.org/tracker/CVE-2024-24680
https://github.com/django/django/commit/572ea07e84b38ea8de0551f4b4eda685d91d09d2
https://docs.djangoproject.com/en/dev/releases/4.2.10/

(LOCAL REV: NOT UPSTREAM) -- Not applicable, upstream out of maintenance

Signed-off-by: Rahul Janani Pandi <RahulJanani.Pandi@windriver.com>