postgresql: Fix CVE-2024-0985
Issue: LIN1023-3631

Late privilege drop in REFRESH MATERIALIZED VIEW CONCURRENTLY in PostgreSQL
allows an object creator to execute arbitrary SQL functions as the command
issuer. The command intends to run SQL functions as the owner of the
materialized view, enabling safe refresh of untrusted materialized views.
The victim is a superuser or member of one of the attacker's roles. The
attack requires luring the victim into running REFRESH MATERIALIZED VIEW
CONCURRENTLY on the attacker's materialized view. As part of exploiting
this vulnerability, the attacker creates functions that use CREATE RULE to
convert the internally-built temporary table to a view. Versions before
PostgreSQL 15.6, 14.11, 13.14, and 12.18 are affected. The only known
exploit does not work in PostgreSQL 16 and later. For defense in depth,
PostgreSQL 16.2 adds the protections that older branches are using to fix
their vulnerability.

Upstream commit have wrong CVE mentioned but the correct one is CVE-2024-0985

References:
https://nvd.nist.gov/vuln/detail/CVE-2024-0985
https://security-tracker.debian.org/tracker/CVE-2024-0985

(LOCAL REV: NOT UPSTREAM) -- Not applicable, upstream out of maintenance

Signed-off-by: Soumya Sambu <soumya.sambu@windriver.com>